refactor and rename + checkUnauthorized
