+++ src/__tests__/auth/Login.reducer.spec.ts
@@ -1,49 +1,49 @@
 import store from "../../app/common/store";
 import {
-  receiveLoginFailure,
-  receiveLoginSuccess,
-  requestLogin,
-} from "../../app/Login/Login.slice";
+	receiveLoginFailure,
+	receiveLoginSuccess,
+	requestLogin,
+} from "../../pages/Login/LoginPage.slice";
 
 describe("Auth Reducer", () => {
-  it("should handle request login", () => {
-    const baseLoginState = store.getState().login;
+	it("should handle request login", () => {
+		const baseLoginState = store.getState().login;
 
-    store.dispatch(requestLogin({ username: "test", password: "test" }));
+		store.dispatch(requestLogin({ username: "test", password: "test" }));
 
-    const expected = {
-      ...baseLoginState,
-      errorMessage: "",
-      isLoading: true,
-    };
-    expect(store.getState().login).toEqual(expected);
-  });
+		const expected = {
+			...baseLoginState,
+			errorMessage: "",
+			isLoading: true,
+		};
+		expect(store.getState().login).toEqual(expected);
+	});
 
-  it("should handle receive login success", () => {
-    const baseLoginState = store.getState().login;
+	it("should handle receive login success", () => {
+		const baseLoginState = store.getState().login;
 
-    const token = "token";
-    store.dispatch(receiveLoginSuccess(token));
+		const token = "token";
+		store.dispatch(receiveLoginSuccess(token));
 
-    const expected = {
-      ...baseLoginState,
-      errorMessage: "",
-      isLoading: false,
-    };
-    expect(store.getState().login).toEqual(expected);
-  });
+		const expected = {
+			...baseLoginState,
+			errorMessage: "",
+			isLoading: false,
+		};
+		expect(store.getState().login).toEqual(expected);
+	});
 
-  it("should handle receive login success", () => {
-    const baseLoginState = store.getState().login;
+	it("should handle receive login success", () => {
+		const baseLoginState = store.getState().login;
 
-    const msg = "test-error";
-    store.dispatch(receiveLoginFailure(msg));
+		const msg = "test-error";
+		store.dispatch(receiveLoginFailure(msg));
 
-    const expected = {
-      ...baseLoginState,
-      errorMessage: msg,
-      isLoading: false,
-    };
-    expect(store.getState().login).toEqual(expected);
-  });
+		const expected = {
+			...baseLoginState,
+			errorMessage: msg,
+			isLoading: false,
+		};
+		expect(store.getState().login).toEqual(expected);
+	});
 });
