+++ src/app/common/rootReducer.ts
@@ -1,14 +1,14 @@
 import { combineReducers } from "redux";
 import problemEditorContainerReducer from "../../pages/ProblemEditor/ProblemEditorContainer/problemEditorContainerSlice";
-import { loginSlice } from "../Login/Login.slice";
+import { loginSlice } from "../../pages/Login/LoginPage.slice";
 /* import slices of state here */
 
 /* Place the object keys for state here followed by the reducer taken from that slice
     e.g. key : keyReducer
 */
 const rootReducer = combineReducers({
-  login: loginSlice.reducer,
-  problemEditorContainer: problemEditorContainerReducer,
+	login: loginSlice.reducer,
+	problemEditorContainer: problemEditorContainerReducer,
 });
 
 export default rootReducer;
