+++ src/app/common/rootSaga.ts
@@ -1,7 +1,7 @@
 import { all } from "redux-saga/effects";
-import loginSaga from "../Login/Login.sagas";
+import loginSaga from "../../pages/Login/LoginPage.sagas";
 
 /** Place additional saga functions in the array */
 export default function* rootSaga() {
-  yield all([loginSaga()]);
+	yield all([loginSaga()]);
 }
